refactor(Prompts): move sidebar components to sidebar/ subdirectory
