+++ client/src/components/Prompts/sidebar/PanelNavigation.tsx
