+++ client/src/components/Prompts/sidebar/index.ts
@@ -1,0 +1,4 @@
+export { default as GroupSidePanel } from './GroupSidePanel';
+export { default as PromptsAccordion } from './PromptsAccordion';
+export { default as FilterPrompts } from './FilterPrompts';
+export { default as PanelNavigation } from './PanelNavigation';
